requirements.txt
-i https://pypi.org/simple
alembic==1.12.1 ; python_version >= '3.7'
aniso8601==9.0.1
bcrypt==4.0.1 ; python_version >= '3.6'
blinker==1.7.0 ; python_version >= '3.8'
click==8.1.7 ; python_version >= '3.7'
flask==3.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.1 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
honcho==1.1.0
importlib-metadata==6.8.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.3.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
packaging==23.2 ; python_version >= '3.7'
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23
sqlalchemy-serializer==1.4.1
typing-extensions==4.8.0 ; python_version >= '3.8'
werkzeug==3.0.1 ; python_version >= '3.8'
zipp==3.17.0 ; python_version >= '3.8'
